Split MyApp widgets, add repositories and sign page route
