+++ lib/main.dart
@@ -1,15 +1,21 @@
 import 'package:amadeo/helpers/net_helper.dart';
-import 'package:amadeo/home_screen.dart';
-import 'package:amadeo/pages/export.dart';
-import 'package:amadeo/utils/style.dart';
+import 'package:amadeo/my_app.dart';
+import 'package:amadeo/repositories/document_repository.dart';
+import 'package:amadeo/repositories/sdn_selected_repository.dart';
+import 'package:amadeo/simple_bloc_delegate.dart';
 import 'package:commercio_ui/commercio_ui.dart';
 import 'package:commercio_ui/core/utils/export.dart';
 import 'package:commerciosdk/export.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
+
+  if (!kReleaseMode) {
+    BlocSupervisor.delegate = SimpleBlocDelegate();
+  }
 
   final commercioAccount = StatefulCommercioAccount(
     networkInfo: NetworkInfo(
@@ -21,88 +27,31 @@ void main() {
       lcdUrl: ChainNet.dev.lcdUrl,
     ),
   );
-  final commercioDocs =
-      StatefulCommercioDocs(commercioAccount: commercioAccount);
-  final commercioId = StatefulCommercioId(commercioAccount: commercioAccount);
-  final commercioMint =
-      StatefulCommercioMint(commercioAccount: commercioAccount);
-  final commercioMembership =
-      StatefulCommercioMembership(commercioAccount: commercioAccount);
-
-  final providers = [
-    BlocProvider(
-      create: (_) => CommercioAccountBloc(commercioAccount: commercioAccount),
-    ),
-    BlocProvider(
-      create: (_) => CommercioIdBloc(commercioId: commercioId),
-    ),
-    BlocProvider(
-      create: (_) => CommercioDocsBloc(
-        commercioDocs: commercioDocs,
-        commercioId: commercioId,
-      ),
-    ),
-    BlocProvider(
-      create: (_) => CommercioMintBloc(commercioMint: commercioMint),
-    ),
-    BlocProvider(
-      create: (_) =>
-          CommercioMembershipBloc(commercioMembership: commercioMembership),
-    ),
-    BlocProvider<CommercioDocsEncDataBloc>(
-      create: (_) => CommercioDocsEncDataBloc(),
-    ),
-  ];
 
   runApp(
-    MultiBlocProvider(
-      providers: providers,
-      child: RepositoryProvider<StatefulCommercioAccount>.value(
-        value: commercioAccount,
-        child: MyApp(),
-      ),
+    MultiRepositoryProvider(
+      providers: [
+        RepositoryProvider.value(value: commercioAccount),
+        RepositoryProvider(
+          create: (_) =>
+              StatefulCommercioDocs(commercioAccount: commercioAccount),
+        ),
+        RepositoryProvider(
+          create: (_) =>
+              StatefulCommercioId(commercioAccount: commercioAccount),
+        ),
+        RepositoryProvider(
+          create: (_) =>
+              StatefulCommercioMint(commercioAccount: commercioAccount),
+        ),
+        RepositoryProvider(
+          create: (_) =>
+              StatefulCommercioMembership(commercioAccount: commercioAccount),
+        ),
+        RepositoryProvider(create: (_) => DocumentRepository()),
+        RepositoryProvider(create: (_) => SdnSelectedDataRepository()),
+      ],
+      child: const MyApp(),
     ),
   );
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Amadeo',
-      theme: companyTheme,
-      initialRoute: '/',
-      routes: {
-        '/': (_) => const HomeScreen(),
-        '/1-account': (_) => const CommercioAccountPage(),
-        '/1-account/generate-new-wallet': (_) => const GenerateNewWalletPage(),
-        '/1-account/restore-wallet-from-mnemonic': (_) =>
-            const RestoreWalletFromMnemonicPage(),
-        '/1-account/restore-wallet-from-secure-storage': (_) =>
-            const RestoreWalletFromSecureStoragePage(),
-        '/1-account/share-qr-code': (_) => const ShareQRCodePage(),
-        '/1-account/request-invite-free-tokens': (_) =>
-            const RequestInviteFreeTokensPage(),
-        '/1-account/check-account-balance': (_) =>
-            const CheckAccountBalancePage(),
-        '/1-account/send-tokens': (_) => const SendTokensPage(),
-        '/1-account/generate-many-addresses': (_) =>
-            const GenerateManyAddressesPage(),
-        '/2-id': (_) => const CommercioIdPage(),
-        '/2-id/create-ddo': (_) => const CreateDDOPage(),
-        '/2-id/request-powerup': (_) => const RequestPowerupPage(),
-        '/3-docs': (_) => const CommercioDocsPage(),
-        '/3-docs/share-doc': (_) => const ShareDocPage(),
-        '/3-docs/send-receipt': (_) => const SendReceiptPage(),
-        '/3-docs/document-list': (_) => const DocumentListPage(),
-        '/3-docs/receipt-list': (_) => const ReceiptListPage(),
-        '/5-mint': (_) => const CommercioMintPage(),
-        '/5-mint/open-cdp': (_) => const OpenCdpPage(),
-        '/5-mint/close-cdp': (_) => const CloseCdpPage(),
-        '/6-kyc': (_) => const CommercioKYCPage(),
-        '/6-kyc/buy-membership': (_) => const BuyMembershipPage(),
-        '/6-kyc/invite-member': (_) => const InviteMemberPage(),
-      },
-    );
-  }
 }
